:lipstick: Fix broken selectors.
Changed pseudo classes on selectors for first and last articles in
the page from :first-child to :first-of-type and :last-child to
:last-of-type.

This is due to :first-child breaking because a h2 was added to the
section that the articles are contained in (see previous commit).

Since the articles are not the only children in the section anymore
the selector broke. The :last-child selector did not break but it
seemed like good practice to change it as something could be added
to the bottom of the section in the markup later on e.g the footer.