:wheelchair: Update css to fix accessibility issues.
Headings elements added in the following commit:

59c8781b88fc2cdf94db43f3ea9918006c152f37

were hidden to conform to the design.

They were added for the accessibility purposes. The problem was
that screen readers can't see elements when using display: none;

The styling has been updated in this commit so that the heading
elements can better serve their purpose.